Refactor NoSuchSchemaException
