+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NoSuchSchemaException.java
@@ -1,0 +1,13 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class NoSuchSchemaException extends NakadiRuntimeBaseException {
+
+    public NoSuchSchemaException(final String message) {
+        super(message);
+    }
+
+    public NoSuchSchemaException(final String msg, final Exception cause) {
+        super(msg, cause);
+    }
+
+}
